backend/requirements.txt
fastapi==0.118.0
uvicorn[standard]==0.37.0
sqlalchemy==2.0.43
psycopg[binary]==3.2.10
python-multipart==0.0.20
pydantic-settings==2.10.1
httpx==0.28.1
faster-whisper==1.1.1
